Separate out hiding listings logic into its own function
