+++ src/utils/addHideBtnToPage.ts
@@ -1,6 +1,5 @@
-import { HIDE_BUTTON_TEXTCONTENT, LISTING_ID_ATTRIBUTE_NAME, searchListingEleQuery } from "../constants.js"; 
+import { HIDE_BUTTON_TEXTCONTENT, LISTING_ID_ATTRIBUTE_NAME } from "../constants.js"; 
 import { createElement } from "./createElement.js";
-import { getAll } from "./database/chromeStorageLocalDb.js";
 import { getHideButtonId } from "./getHideButtonId.js";
 import { getSearchListings } from "./getSearchListings.js";
 import { onclickHideButton } from "./onclickHideButton.js";
@@ -25,17 +24,6 @@ export function addHideBtnToPage() {
                     }
                 )
                 l.element.appendChild(btnEle)
-            }
-        })
-    }, 250)
-
-    //@ts-ignore
-    const removeListingInterval = SetIntervalAsync.dynamic.setIntervalAsync(async () => {
-        const hiddenListings = await getAll('HiddenListing')
-        const listings = getSearchListings()
-        listings?.forEach(l => {
-            if (hiddenListings.find(hidden => hidden.id === l.listingId)) {
-                l.element.hidden = true
             }
         })
     }, 250)
